+++ react/GoToCheckout.tsx
@@ -1,0 +1,21 @@
+import React, { FunctionComponent } from 'react'
+import { FormattedMessage } from 'react-intl'
+import { Button } from 'vtex.styleguide'
+
+const GoToCheckoutButton: FunctionComponent = () => {
+  return (
+    <div className="pv3 bg-white">
+      <Button
+        id="proceed-to-checkout"
+        href="/checkout/#payment"
+        variation="primary"
+        size="large"
+        block
+      >
+        <FormattedMessage id="store/cart.checkout" />
+      </Button>
+    </div>
+  )
+}
+
+export default GoToCheckoutButton