+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- "Go to checkout" button is now a sticky footer on small screens.
+
 ## [0.13.2] - 2019-10-15
 
 ### Fixed